Adapt processing to Catalina
It looks like the log format has slightly changed with macOS Catalina.
The file-issue-extension arguments lack the spaces after the name of the
argument and network operations lack a space before the argument itself.
The first will lead to the custom sandbox check implementation not being
able to parse the argument properly and the latter will lead to the
operation name including the argument, which is then passed to
sandbox_check and fails.

The changes process the log in a backward compatible manner.