Prepend slash to $_HOME$ placeholder
There is a home-path-ancestors function defined in the application.sb
base profile, which in turn uses path-ancestors from the internal SBPL
definition. This function performs a check, whether a path starts with a
slash. Since the function is never called in the application base
profile, this usually does not cause issues. However, Safari utilizes
the com.apple.security.temporary-exception.sbpl entitlement and the
function is called during evaluation, e. g., with simbple. When we pass
a normalised profile, simbple will crash due to invalid handling when
outputting errors – unless it is compiled with SBPL_DEBUG=1. The issue
is that for normalised profiles, the user's home directory gets replaced
with a $_HOME$ placeholder. The path-ancestors function however, checks
whether a path starts with a slash, but our placeholder starts with a
dollar sign. This triggers an error [1]. In order to avoid this, we
simply prepend a slash in front of the placeholder in that case.

[1] https://github.com/0xbf00/simbple/blob/cf9556aebb14eaa215a2b0510d5c64d9ce103243/src/scm/sbpl_v1.scm#L222