Improve rule mapping / generalisation of results
Instead of transforming the results directly, we now simply calculate a
rule mapping, which can be used for mapping original rules to
normalised or generalised ones. The results will always use the original
rule indexes. Instead of returning -1 for unmatched rules, we simply
omit them from the result. This way the mapping is simply a mapping from
integer to integer and not some nested list or union structure.

In addition, the mapping itself was improved by respecting replacements
made during normalisation. Instead of only comparing the action and
operation of rules, we revert normalisation and match the full rule.
Additionally, it turned out that the loop was aborted too early and some
rules got not mapped correctly.