Improve matching
Sorry, this is quite a big change.

First, I moved the profile reduction and match result interpretation to
Python. The matcher utility that is still written in C/C++ now only
takes a profile and processed logs and returns a list of match results
for each processed log entry. That way, we still have batch processing
per profile.

Second, the matching now handles errors more explicitly. I think it is
better to run into an error than to get invalid results. A re-checked
rule was considered denied if an error occurred, however, this is not
always the case, e.g., if we simply do not test the rule due to a
missing customized check.

Third, in addition to reducing the profile, we now mutate every tested
rule (last rule in the profile), i. e., transform an allow to a deny
and vice versa. Since, twice as many profiles now need to be tested,
matching now takes twice as long, however, the results are worth it.
That way we can detect rules that would change the outcome, if the
action was different, meaning that the rule was redundant for the given
log entry. This does not necessarily mean that the rule is redundant in
itself, as it can still be a deciding rule for other log entries. If we
can identify a deciding rule for a log entry, even an implicit default,
we can even pinpoint the rules to which another rule was redundant. This
might point to superfluous rules that could be simplified.

Fourth, for simplicity we include a list of log entries that could not
be matched. These might need futher investigation and proabably a custom
sandbox_check implementation.