Update IOKit mappings (ref #2)
I added a utility that utilizes Siguza's ioscan in order to calculate a
mapping between `IOService`s and their respective `IOUserClient`s. The
utility can generate the initializer for our mapping struct in
`iokit.c`.

I archived the mappings for known versions of macOS. Maybe we want to
use the JSON files in the future.

This also fixes the mappings I "repaired" erroneously:
- f937d96fda155c8cf0657c60de1aff01f44592cf
- 3256bb4539c1e516d58ae882924927fbae29fb09