Improve checking shared memory
As we perform checks after an application has already quit, we might try
to access shared memory objects, which are already gone. Therefore, we
now try to create and open the object. If creation does not succeed we
do now know whether opening is allowed or not and an error will be
returned, however if creating is allowed as well, we can check the open
operation.