Return unknown decision for non-existing shared memory objects
We first try to create them, but if creation is denied, the sandbox
could still allow reading or writing to the objects. Since the object is
cone, we cannot determine a correct result. However, we do not want to
error out in that case, as we cannot repair it anyway.